Users,Roles,Complains,Profile,User Logs, Holidays, Permissions and Suggestions Update By Muhammad Irfan.

+++ routes/roles/roles.php
@@ -4,9 +4,9 @@
 
         Route::get('add' , function (){
             return view('roles.add');
-        })->name('add');
+        })->name('add')->middleware('permission:add role,add');
 
-        Route::post('add','RoleController@store')->name('add');
+        Route::post('add','RoleController@store')->name('add')->middleware('permission:add role,add');
 //        Route::post('add', ['as' => '.add' , 'uses' => 'RoleController@store' ]);
 
         /*Route::get('show' , function (){
@@ -17,12 +17,12 @@
 
         Route::get('Edit' , function (){
             return view('roles.Edit');
-        })->name('Edit');
+        })->name('Edit')->middleware('permission:edit role,Edit');
 
-        Route::get('Edit/{id}','RoleController@edit')->name('Edit');
-        Route::post('Edit/{id}','RoleController@update')->name('Edit');
+        Route::get('Edit/{id}','RoleController@edit')->name('Edit')->middleware('permission:edit role,Edit');
+        Route::post('Edit/{id}','RoleController@update')->name('Edit')->middleware('permission:edit role,Edit');
 
-        Route::delete('delete/{id}','RoleController@destroy')->name('delete');
+        Route::delete('delete/{id}','RoleController@destroy')->name('delete')->middleware('permission:delete role,delete');
 
     });
